- Refactoring to support Docker pipeline execution
- Better handling of test execution timeout
- Extends tests for utils
- Adds a Docker-specific gradle task
- Better gitignore
