+++ src/python/test/pipelines_utils/test_utils.py
@@ -12,4 +12,6 @@ class UtilsTestCase(unittest.TestCase):
         self.assertEquals(utils.round_sig(1.23456789, 3), 1.23)
         self.assertEquals(utils.round_sig(1.23456789, 6), 1.23457)
 
+        self.assertEquals(utils.round_sig(0.00456789, 3), 0.00457)
+
         self.assertEquals(utils.round_sig(-1.23456789, 6), -1.23457)
